翻译一段.txt
While this gure might well be dierent in today's environment, the
basic observation is still true, and may even have been aggravated by
the use of instantaneous global communication. Even worse, more
users not only nd more actual bugs, but also articulate more wishes
in general. Be it a genuine error, a feature request, or merely a fun-
damental misunderstanding of the software's operation, the typical
user request is far from being a technically precise bug report. And
each request demands the attention of the developers, consuming
precious time that is not available to actually write code.